Implement easter egg background.
Clicking on the clover send the user to /toggle_easter_egg.
This sets a session variable "easter-egg-mode". The user
is then sent back to the path they came from or the index.

If that flag is enabled, the standard stylesheet is replaced
by the easter egg stylesheet, and <main> uses the "create-bg"
class.